fix: ./json directory (rename to file)
